reduce code duplication for paperclip setup
