+++ paperweight-core/src/main/kotlin/PaperweightCore.kt
@@ -162,20 +162,7 @@ class PaperweightCore : Plugin<Project> {
                 mcVersion.set(target.ext.minecraftVersion)
             }
 
-            paperclipJar {
-                with(target.tasks.named("jar", Jar::class).get())
-
-                val paperclipConfig = target.configurations.named(PAPERCLIP_CONFIG)
-                dependsOn(paperclipConfig, generatePaperclipPatch)
-
-                val paperclipZip = target.zipTree(paperclipConfig.map { it.singleFile })
-                from(paperclipZip) {
-                    exclude("META-INF/MANIFEST.MF")
-                }
-                from(target.zipTree(generatePaperclipPatch.flatMap { it.outputZip }))
-
-                manifest.from(paperclipZip.matching { include("META-INF/MANIFEST.MF") }.elements.map { it.single() })
-            }
+            paperclipJar.configurePaperclipJar(target, generatePaperclipPatch)
         }
     }
